refactor:
- переместил student макеты в предметную папку
- поменял валидацию в хуке useHomeworksByTeachingAssignment
- починил метод теста createGrade_ShouldReturn200
